git-svn-id: svn://svn.ijw.co.nz/svn/OpenRa@1231 993157c7-ee19-0410-b2c4-bb4e9862e678

+++ OpenRa.Game/Region.cs
@@ -20,7 +20,7 @@ namespace OpenRa.Game
 
 		public void Draw(Renderer renderer, Viewport viewport)
 		{
-			renderer.Device.EnableScissor(location.X, location.Y, size.Width, size.Height);
+			renderer.Device.EnableScissor((int)location.X, (int)location.Y, size.Width, size.Height);
 			drawFunction(renderer, viewport);
 			renderer.Device.DisableScissor();
 		}
